Task 11

+++ 0x05-pointers_arrays_strings/101-keygen.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-long int num;
+int num;
 
 srand(time(0));
 num = rand();
